Refactor: Add InfiniteScrollList component
